Add support for cppflags
This is similar to the `cflags` family of compiler flags, but is meant
to be used with the pre-processor in the case where pre-processing needs
to be performed as a separate build command. Alternatively, `cppflags`
can be adapted for use during a post-processing step (i.e. lint).

As with `cflags`, `cppflags` also supports language specific flags:
  cppflags_c
  cppflags_cc
  cppflags_objc
  cppflags_objcc