Add support for specifying tool source extensions
This allows tool specific source extensions to be customized instead of
being limited to the built-in defaults for each tool type. For example,
an assembler tool can specify `arm` extensions as assembly files by
setting `source_extensions = [ "arm" ]` in the `asm` tool. This is
particularly useful for cross platform toolchains, which often recognize
different source extensions as assembly files, where the extension name
is based on the target architecture or platform.