Output dir outside root uses abs paths for sources
Forces path output of source files to be in absolute paths instead of
relative paths whenever the output directory is outside of the source
root directory. As a result, source files and other target inputs will
be written out to ninja files using their absolute paths, instead of the
relative path from the build directory. This is desirable in cases where
you would want all source files to be specified with absolute paths on
the command-line used to invoke the compiler. For example, in order to
line up the debug source file paths with a source distribution's install
path, without using any special compiler flags or debugger options to
remap the source path. This becomes even more important for toolchains
that do not support source path remapping at compile time.