Do not use rspfile for gcc link tools
We cannot use response files for gcc linker tools, otherwise the linker
will not set the rpath in the final linked executable. This appears to
be an undocumented feature or bug in the gnu tools. Using a @rspfile
will result in a linked executable without rpath set. However, simply
replacing the @rspfile with the exact contents of the rspfile will
result in a linked executable with rpath set.