Fix errors when root build dir equals source root
When the root build directory is the same as the source root directory,
all source and input files generate "Input to target not generated by a
dependency" error. In such an even, we should check that input/source
files are not in the obj/gen output directories, instead of the root
build directory. Otherwise, just do the normal check that input/source
files are not in the root build directory.