Fix configs not appending precompiled headers
If a precompiled header was specified in a subconfig, the resulting
composite config would not add the precompiled header, even when there
were no other precompiled headers in any other subconfigs or the parent
config. This fix appends the precompiled header from a subconfig if the
composite config does not already have a precompiled header.