Fix help message title for GN commands
For GN commands (i.e. args, desc, ls, refs, etc.), the long help message
was missing the short description of the command, using the usage line
as the help message title.  Hence, causing the markdown output to omit
parts of the usage line.

This change adds a short description for the command as the help message
title, and moves the usage line to the body of the help.