Fix target defaults for templates
When using `set_defaults()` on a template, default values are only
applied when the template is invoked directly using the template name.
However, when the template is invoked using the `target()` function,
default values are not applied.

For example:

  template("my_test") {
    executable(target_name) {
      forward_variables_from(invoker, "*")
    }
  }

  set_defaults("my_test") {
    deps = ["//utils/test_main"]
  }

  my_test("test1") {
    sources = [
      "test1_unittest.cc"
    ]
    # deps are copied from set_defaults("my_test")
  }

  target("my_test", "test2") {
    sources = [
      "test2_unittest.cc"
    ]
    # deps are NOT copied from set_defaults("my_test")
  }

This occurs because the invoked template retrieves default values using
the invoking function's name, which resolves to the template name when
invoked directly, but resolves to "target" when invoked through the
`target()` function. The invoked template must retrieve default values
using the name obtained from its own function definition.