Change ordering of config flags and values
This changes the ordering of config flags and values to better match was
is expected behavior of precedence when specifying flags and values
directly in a target versus a config versus a config inherited from a
dependency.

The current ordering is as follows:

  1. Those set on the current target (not in a config).
  2. Those set on the "configs" on the target in order that the
     configs appear in the list.
  3. Those set on the "all_dependent_configs" on the target in order
     that the configs appear in the list.
  4. Those set on the "public_configs" on the target in order that
     those configs appear in the list.
  5. all_dependent_configs pulled from dependencies, in the order of
     the "deps" list. This is done recursively. If a config appears
     more than once, only the first occurance will be used.
  6. public_configs pulled from dependencies, in the order of the
     "deps" list. If a dependency is public, they will be applied
     recursively.

This makes sense when thought of as an order of precedence for flags and
values to be applied. However, the order on the command-line is not
necessarily the same as the order of precedence. For directories and
libraries, the order they appear on the command-line is the order in
which they are applied, and values that occur earlier on the command-
line take precedence over values that occur later. Though, for generic
compile flags, like optimization, the opposite holds true. In these
cases, flags occurring later on the command-line take precedence, or may
override, flags occurring earlier on the command-line.

Hence, for flags and defines the ordering should be as follows:

  1. public_configs pulled from dependencies, in the order of the
     "deps" list. If a dependency is public, they will be applied
     recursively.
  2. all_dependent_configs pulled from dependencies, in the order of
     the "deps" list. This is done recursively. If a config appears
     more than once, only the first occurance will be used.
  3. Those set on the "public_configs" on the target in order that
     those configs appear in the list.
  4. Those set on the "all_dependent_configs" on the target in order
     that the configs appear in the list.
  5. Those set on the "configs" on the target in order that the
     configs appear in the list.
  6. Those set on the current target (not in a config).

As a result, if an optimization level of -O1 is applied directly on the
current target via the "cflags" variable, then -O1 will take precedence
over any optimization flags pulled from dependencies as well as any
optimization flags set on the "*configs" on the target.