Fix library link order for complete static libs
This fixes the link line when complete static libraries are used such
that dependencies that are lower in the dependency tree occur before
their parents. This is a bug specific to the use of complete static
libraries. When a complete static library has dependencies that are
"final" output (i.e. complete static library, shared library, etc.),
these "final" output dependencies were added in the reverse order,
resulting in unresolved symbols on the final link line.