Fix ordering of config flags and values
This fixes the ordering of config flags and values, such that the flags
and values set on the "*configs" of a target are applied in the order
that the configs appear in the list, instead of the reverse order.
Additionally, for "*configs" pulled from dependencies, they are applied
in the reverse order of the "deps" list, such that config flags and
values for dependencies occurring further down the "deps" list will take
precedence over those occurring earlier in the "deps" list.

Hence, for flags and defines the ordering should be as follows:

  1. public_configs pulled from dependencies, in the reverse order of
     the "deps" list. If a dependency is public, they will be applied
     recursively.
  2. all_dependent_configs pulled from dependencies, in the reverse
     order of the "deps" list. This is done recursively. If a config
     appears more than once, only the first occurance will be used.
  3. Those set on the "public_configs" on the target in order that
     those configs appear in the list.
  4. Those set on the "all_dependent_configs" on the target in order
     that the configs appear in the list.
  5. Those set on the "configs" on the target in order that the
     configs appear in the list.
  6. Those set on the current target (not in a config).