Fix auto generation after `args` command
When the `gn args` command automatically invokes `gn gen` after the
editor has exited, it does so without any of the command-line options
that may have been present when `gn gen` was first called by the user.
This presents a problem in cases when additional command-line options to
`gn gen` must be retained each time Ninja files are regenerated. For
example, if you use IDE options to keep your IDE project files up to
date.

This changes the `gn args` command to automatically update the
`build.ninja.d` dependency file after editing the `args.gn` file, such
that the `args.gn` file is added to the dependecy list if it is missing,
instead of automatically calling `gn gen` to regenerate Ninja files.
This ensures that Ninja marks the build as dirty if the arg file was
created or updated. Hence, Ninja will automatically run the generator
command, if necessary, the next time it is invoked. Futhermore, the
Ninja generator rule ensures that the same command-line arguments
originally used to invoke `gn gen` by the user are retained.