Ensure toolchains are loaded before build files
When the root build file is loaded, it first loads the build config
file, but does not schedule the default toolchain to be loaded. The
default toolchain is instead loaded at some point in the future, instead
of before the root build file. This updates the load order, such that
when the default build config file is loaded, it will automatically
schedule the default toolchain to be loaded before the root build file.