Support binary targets for `get_target_outputs`
This updates the `get_target_outputs` function to support binary
targets. Executables and libraries will return a list of the resulting
binary file(s). The "main output" (the actual binary or library) will
always be the 0th element in the result. Source sets will return a list
of object files which are the outputs from compilation of each source
file.