Set the target toolchain when target is created
In order to add support for binary targets to `get_target_outputs` we
need to set the target toolchain when the target is created instead of
when the target is defined. Doing so when the target is defined is too
late, since `get_target_outputs` may be called before the desired target
is defined. We can set the toolchain in the settings object that is
passed to the target when the target is created. Since toolchains are
always loaded before any build files, we can associate a record's
settings object with a toolchain as soon as the toolchain is loaded.

This also verifies the toolchain has the needed tools when the target is
defined or resolved.