Workaround for glibc bug with external malloc
This implements a simple workaround for a bug in glibc which prevents
the use of statically linked external malloc. Basically, in glibc 2.24
fork() calls internal functions specific to glib's memory allocator.
This workaround simply provides empty implementations for these internal
functions so that the linker drops the need for malloc.o from glibc.